@@ -62,35 +62,35 @@ limitations under the License.
   </developers>
 
   <properties>
-    <java.version>1.8</java.version>
+    <maven.compiler.release>17</maven.compiler.release>
 
     <!-- Plugin version -->
-    <plugin.clean.version>3.1.0</plugin.clean.version>
-    <plugin.install.version>3.0.0-M1</plugin.install.version>
-    <plugin.resources.version>3.2.0</plugin.resources.version>
-    <plugin.surefire.version>3.0.0-M5</plugin.surefire.version>
-    <plugin.compiler.version>3.8.0</plugin.compiler.version>
-    <plugin.jar.version>3.1.1</plugin.jar.version>
-    <plugin.site.version>3.9.1</plugin.site.version>
+    <plugin.clean.version>3.2.0</plugin.clean.version>
+    <plugin.install.version>3.1.1</plugin.install.version>
+    <plugin.resources.version>3.3.1</plugin.resources.version>
+    <plugin.surefire.version>3.1.0</plugin.surefire.version>
+    <plugin.compiler.version>3.11.0</plugin.compiler.version>
+    <plugin.jar.version>3.3.0</plugin.jar.version>
+    <plugin.site.version>4.0.0-M8</plugin.site.version>
     <plugin.github-site.version>0.12</plugin.github-site.version>
-    <plugin.nexus-staging.version>1.6.8</plugin.nexus-staging.version>
-    <plugin.deploy.version>3.0.0-M1</plugin.deploy.version>
-    <plugin.project-info-reports.version>3.1.1</plugin.project-info-reports.version>
-    <plugin.jgitflow.version>1.0-m5.1</plugin.jgitflow.version>
-    <plugin.gpg.version>1.6</plugin.gpg.version>
-    <plugin.source.version>3.0.1</plugin.source.version>
-    <plugin.javadoc.version>3.1.0</plugin.javadoc.version>
-    <plugin.checkstyle.version>3.1.2</plugin.checkstyle.version>
+    <plugin.nexus-staging.version>1.6.13</plugin.nexus-staging.version>
+    <plugin.deploy.version>3.1.1</plugin.deploy.version>
+    <plugin.project-info-reports.version>3.4.3</plugin.project-info-reports.version>
+    <plugin.jgitflow.version>1.0-m10</plugin.jgitflow.version>
+    <plugin.gpg.version>3.1.0</plugin.gpg.version>
+    <plugin.source.version>3.2.1</plugin.source.version>
+    <plugin.javadoc.version>3.5.0</plugin.javadoc.version>
+    <plugin.checkstyle.version>3.2.2</plugin.checkstyle.version>
     <plugin.github-release.version>1.4.0</plugin.github-release.version>
-    <plugin.jxr.version>3.0.0</plugin.jxr.version>
-    <plugin.pmd.version>3.14.0</plugin.pmd.version>
-    <plugin.enforcer.version>3.0.0-M3</plugin.enforcer.version>
+    <plugin.jxr.version>3.3.0</plugin.jxr.version>
+    <plugin.pmd.version>3.21.0</plugin.pmd.version>
+    <plugin.enforcer.version>3.3.0</plugin.enforcer.version>
 
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
-    <org.springframework.boot.version>2.3.8.RELEASE</org.springframework.boot.version>
-    <org.slf4j.version>1.7.30</org.slf4j.version>
+    <org.springframework.boot.version>3.0.0</org.springframework.boot.version>
+    <org.slf4j.version>2.0.7</org.slf4j.version>
   </properties>
 
   <distributionManagement>
