@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Joao Silva
+ * Copyright 2018-2023 Joao Silva
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,31 +16,84 @@
 
 package com.github.jpmsilva.jsystemd;
 
+import static com.github.jpmsilva.jsystemd.SystemdUtilities.hasNotifySocket;
+import static com.github.jpmsilva.jsystemd.SystemdUtilities.isLinux;
+import static java.lang.invoke.MethodHandles.lookup;
+import static java.util.Objects.requireNonNull;
+import static org.slf4j.LoggerFactory.getLogger;
+
+import java.io.IOException;
+import java.net.StandardProtocolFamily;
+import java.net.UnixDomainSocketAddress;
+import java.nio.ByteBuffer;
+import java.nio.channels.SocketChannel;
+import java.nio.file.Path;
 import org.jetbrains.annotations.NotNull;
+import org.slf4j.Logger;
 
 /**
- * Interface of all systemd integration library implementations.
+ * Low level API that interfaces with systemd through a {@link UnixDomainSocketAddress}. Not meant for direct usage.
  *
  * @author Joao Silva
- * @see <a href="https://www.freedesktop.org/software/systemd/man/sd_notify.html">sd_notify specification</a>
+ * @see <a href="https://www.freedesktop.org/software/systemd/man/sd_notify.html">sd_notify</a>
  */
-interface SystemdNotify {
+public class SystemdNotify {
 
-  /**
-   * If this library is usable under current process conditions (operating system type, systemd available, etc...).
-   *
-   * @return {@code true} if and only if the library can be used
-   */
-  default boolean usable() {
-    return false;
+  private static final Logger logger = getLogger(lookup().lookupClass());
+
+  private static final SocketChannel channel = createSocketChannel();
+
+  private static SocketChannel createSocketChannel() {
+    if (isLinux() && hasNotifySocket()) {
+      SocketChannel channel = null;
+      try {
+        Path socketPath = SystemdUtilities.notifySocketPath();
+        if (socketPath.toFile().exists()) {
+          channel = SocketChannel.open(StandardProtocolFamily.UNIX);
+          channel.configureBlocking(true);
+          channel.connect(UnixDomainSocketAddress.of(socketPath));
+          channel.shutdownInput();
+          return channel;
+        } else {
+          logger.warn("Specified socket path does not exist - systemd integration disabled: " + socketPath);
+        }
+      } catch (IOException e) {
+        if (channel != null) {
+          try {
+            channel.close();
+          } catch (IOException ex) {
+            e.addSuppressed(ex);
+          }
+        }
+        if (logger.isDebugEnabled()) {
+          logger.warn("Could not connect to systemd socket", e);
+        } else {
+          logger.warn("Could not connect to systemd socket: " + e.getMessage());
+        }
+      }
+    }
+    return null;
   }
 
   /**
-   * Notifies systemd that the service unit has completed startup.
+   * Allows knowing if this library is usable under current execution conditions (operating system type, systemd available, etc...).
+   *
+   * @return {@code true} if and only if the library can be used
+   */
+  static boolean usable() {
+    return channel != null && channel.isConnected();
+  }
+
+  /**
+   * Notifies systemd that the program has completed startup.
    *
    * @see <a href="https://www.freedesktop.org/software/systemd/man/sd_notify.html#READY=1">ready</a>
    */
-  default void ready() {
+  static void ready() {
+    if (usable()) {
+      logger.info("Notifying systemd that service is ready");
+      invoke("READY=1");
+    }
   }
 
   /**
@@ -48,7 +101,11 @@ interface SystemdNotify {
    *
    * @see <a href="https://www.freedesktop.org/software/systemd/man/sd_notify.html#STATUS=%E2%80%A6">status</a>
    */
-  default void status(@NotNull String message) {
+  static void status(@NotNull String message) {
+    if (usable()) {
+      logger.debug("Notifying systemd that service status is {}", requireNonNull(message, "Message must not be null"));
+      invoke("STATUS=" + message);
+    }
   }
 
   /**
@@ -57,7 +114,11 @@ interface SystemdNotify {
    * @see <a href="https://www.freedesktop.org/software/systemd/man/sd_notify.html#EXTEND_TIMEOUT_USEC=%E2%80%A6">extend timeout</a>
    * @see <a href="https://github.com/systemd/systemd/blob/master/NEWS">news</a>
    */
-  default void extendTimeout(long timeout) {
+  static void extendTimeout(long timeout) {
+    if (usable()) {
+      logger.debug("Extending startup timeout with {} microseconds", timeout);
+      invoke("EXTEND_TIMEOUT_USEC=" + timeout);
+    }
   }
 
   /**
@@ -65,6 +126,81 @@ interface SystemdNotify {
    *
    * @see <a href="https://www.freedesktop.org/software/systemd/man/sd_notify.html#WATCHDOG=1">watchdog</a>
    */
-  default void watchdog() {
+  static void watchdog() {
+    if (usable()) {
+      logger.debug("Updating watchdog timestamp");
+      invoke("WATCHDOG=1");
+    }
+  }
+
+  /**
+   * Notifies systemd that the program is stopping.
+   *
+   * @see <a href="https://www.freedesktop.org/software/systemd/man/sd_notify.html#STOPPING=1">stopping</a>
+   */
+  static void stopping() {
+    if (usable()) {
+      logger.info("Notifying systemd that service is stopping");
+      invoke("STOPPING=1");
+    }
+  }
+
+  /**
+   * Low level method that sends the {@code sd_notify} formatted message to systemd.
+   *
+   * @param message the message to send, according to <a href="https://www.freedesktop.org/software/systemd/man/sd_notify.html#Description">specification</a>
+   */
+  private static void invoke(String message) {
+    requireNonNull(message);
+    if (usable() && !message.isEmpty()) {
+      try {
+        synchronized (requireNonNull(channel)) {
+          ByteBuffer buffer = ByteBuffer.allocate(1024);
+          buffer.clear();
+          if (message.endsWith("\n")) {
+            buffer.put(message.getBytes());
+          } else {
+            buffer.put((message + "\n").getBytes());
+          }
+          buffer.flip();
+          while (buffer.hasRemaining()) {
+            channel.write(buffer);
+          }
+        }
+      } catch (IOException e) {
+        if (logger.isDebugEnabled()) {
+          logger.warn("Failed to send systemd message " + message, e);
+        } else {
+          logger.warn("Failed to send systemd message " + message + ": " + e.getMessage());
+        }
+      }
+    }
+  }
+
+  /**
+   * Registers a JVM shutdown hook that closes the system integration channel.
+   *
+   * @see Runtime#addShutdownHook(Thread)
+   * @see #close()
+   */
+  public static void registerShutdownHook() {
+    Runtime.getRuntime().addShutdownHook(new Thread(SystemdNotify::close));
+  }
+
+  /**
+   * Attempts to orderly close the systemd integration channel.
+   *
+   * <p>Normally, the integration channel will be closed when the JVM shuts down.<br>
+   * However, you may wish of explicitly close the integration channel, to ensure that all closeable resources are effectively closed.<br> As such, this method
+   * should only be called at most once during the lifecycle of the JVM.
+   */
+  public static void close() {
+    if (channel != null) {
+      try {
+        channel.close();
+      } catch (IOException ignored) {
+        // This is a best effort at this point
+      }
+    }
   }
 }
