@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Joao Silva
+ * Copyright 2018-2023 Joao Silva
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,21 +16,24 @@
 
 package com.github.jpmsilva.jsystemd;
 
+import static com.github.jpmsilva.jsystemd.SystemdUtilities.isUnderSystemd;
+import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.TimeUnit.MICROSECONDS;
 import static java.util.concurrent.TimeUnit.SECONDS;
 
-import com.github.jpmsilva.jsystemd.SystemdNotifyApplicationRunStatusProvider.ApplicationState;
-import java.util.Objects;
+import com.github.jpmsilva.jsystemd.SystemdApplicationRunStatusProvider.ApplicationState;
+import java.time.Duration;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
+import org.springframework.boot.ConfigurableBootstrapContext;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.SpringApplicationRunListener;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.core.env.ConfigurableEnvironment;
 
 /**
- * A Spring Application Run Listener that sets up a {@link SystemdNotifyApplicationRunStatusProvider} to provide status updates of the current phase of the
+ * A Spring Application Run Listener that sets up a {@link SystemdApplicationRunStatusProvider} to provide status updates of the current phase of the
  * application life cycle.
  *
  * @author Joao Silva
@@ -39,12 +42,12 @@ import org.springframework.core.env.ConfigurableEnvironment;
 public class SystemdSpringApplicationRunListener implements SpringApplicationRunListener {
 
   private static final String SYSTEMD_BEAN_NAME = "systemd";
-  private static final boolean IS_UNDER_SYSTEMD = SystemdUtilities.isUnderSystemd();
-  private static final Systemd SYSTEMD = ensureSystemd();
+
+  private final Systemd systemd = createSystemd();
 
   private final int applicationId;
   @Nullable
-  private SystemdNotifyApplicationRunStatusProvider provider;
+  private SystemdApplicationRunStatusProvider provider;
 
   /**
    * Mandatory constructor of SpringApplicationRunListener.
@@ -54,65 +57,48 @@ public class SystemdSpringApplicationRunListener implements SpringApplicationRun
    */
   @SuppressWarnings({"PMD.UnusedFormalParameter", "unused"})
   public SystemdSpringApplicationRunListener(@NotNull SpringApplication springApplication, String[] args) {
-    applicationId = Objects.requireNonNull(springApplication, "Spring application must not be null").hashCode();
-    if (IS_UNDER_SYSTEMD) {
-      provider = new SystemdNotifyApplicationRunStatusProvider(Objects.requireNonNull(SYSTEMD, "Systemd must not be null"), applicationId);
+    applicationId = requireNonNull(springApplication, "Spring application must not be null").hashCode();
+    if (isUnderSystemd()) {
+      provider = new SystemdApplicationRunStatusProvider(requireNonNull(systemd, "Systemd must not be null"), applicationId);
     }
   }
 
-  private static Systemd ensureSystemd() {
-    if (IS_UNDER_SYSTEMD) {
-      SystemdUtilities.logSystemdStatus();
-      return Systemd.builder()
-          .watchdog(SystemdUtilities.watchdogUsec() / 2, MICROSECONDS)
-          .statusUpdate(5, SECONDS)
-          .build();
+  private static Systemd createSystemd() {
+    if (isUnderSystemd()) {
+      return Systemd.builder().watchdog(SystemdUtilities.watchdogUsec() / 2, MICROSECONDS).statusUpdate(5, SECONDS).build();
     } else {
       return null;
     }
   }
 
-  /**
-   * {@inheritDoc}
-   */
   @Override
-  public void starting() {
+  public void starting(ConfigurableBootstrapContext bootstrapContext) {
     if (provider != null) {
       provider.state(ApplicationState.STARTING);
     }
   }
 
-  /**
-   * {@inheritDoc}
-   */
   @Override
-  public void environmentPrepared(ConfigurableEnvironment environment) {
+  public void environmentPrepared(ConfigurableBootstrapContext bootstrapContext, ConfigurableEnvironment environment) {
     if (provider != null) {
       provider.state(ApplicationState.ENVIRONMENT_PREPARED);
     }
   }
 
-  /**
-   * {@inheritDoc}
-   */
   @Override
   public void contextPrepared(ConfigurableApplicationContext context) {
     if (provider != null) {
       provider.state(ApplicationState.CONTEXT_PREPARED);
 
       ConfigurableListableBeanFactory beanFactory = context.getBeanFactory();
-      Systemd systemd = Objects.requireNonNull(SYSTEMD, "Systemd must not be null");
       if (!beanFactory.containsSingleton(SYSTEMD_BEAN_NAME)) {
-        beanFactory.registerSingleton(SYSTEMD_BEAN_NAME, systemd);
+        beanFactory.registerSingleton(SYSTEMD_BEAN_NAME, requireNonNull(systemd));
       }
       beanFactory.registerSingleton("systemdNotifyApplicationContextStatusProvider",
-          new SystemdNotifyApplicationContextStatusProvider(systemd, applicationId, context.getId(), beanFactory));
+          new SystemdApplicationContextStatusProvider(requireNonNull(systemd), applicationId, context.getId(), beanFactory));
     }
   }
 
-  /**
-   * {@inheritDoc}
-   */
   @Override
   public void contextLoaded(ConfigurableApplicationContext context) {
     if (provider != null) {
@@ -120,24 +106,17 @@ public class SystemdSpringApplicationRunListener implements SpringApplicationRun
     }
   }
 
-  /**
-   * {@inheritDoc}
-   */
   @Override
-  public void started(ConfigurableApplicationContext context) {
+  public void started(ConfigurableApplicationContext context, Duration timeTaken) {
+    if (provider != null) {
+      provider.state(ApplicationState.STARTED, timeTaken);
+    }
   }
 
-  /**
-   * {@inheritDoc}
-   */
   @Override
-  public void running(ConfigurableApplicationContext context) {
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public void failed(ConfigurableApplicationContext context, Throwable exception) {
+  public void ready(ConfigurableApplicationContext context, Duration timeTaken) {
+    if (provider != null) {
+      provider.state(ApplicationState.READY, timeTaken);
+    }
   }
 }
